refactor: created shared/serializer directory and moved all the shared serializer files into it
